Issue 670: Improve UserAccessModel class structure.
