--- strongbox-security/strongbox-security-api/src/main/java/org/carlspring/strongbox/security/Privileges.java
@@ -1,51 +1,0 @@
-package org.carlspring.strongbox.security;
-
-import java.util.LinkedHashSet;
-import java.util.Set;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- * @author mtodorov
- */
-@XmlRootElement(name = "privileges")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class Privileges
-{
-
-    @XmlElement(name = "privilege")
-    private Set<Privilege> privileges = new LinkedHashSet<>();
-
-    public Privileges()
-    {
-    }
-
-    public Privileges(Set<Privilege> privileges)
-    {
-        this.privileges = privileges;
-    }
-
-    public Set<Privilege> getPrivileges()
-    {
-        return privileges;
-    }
-
-    public void setPrivileges(Set<Privilege> privileges)
-    {
-        this.privileges = privileges;
-    }
-
-    @Override
-    public String toString()
-    {
-        final StringBuilder sb = new StringBuilder("\n\tPrivileges{");
-        sb.append("privileges=").append(privileges);
-        sb.append('}');
-        
-        return sb.toString();
-    }
-    
-}
