+++ strongbox-security/strongbox-user-management/src/main/java/org/carlspring/strongbox/config/UsersConfig.java
@@ -135,10 +135,8 @@ public class UsersConfig
         {
             AccessModel internalAccessModel = new AccessModel();
             userAccessModel.getStorages()
-                           .getStorages()
                            .forEach(storage ->
                                             storage.getRepositories()
-                                                   .getRepositories()
                                                    .forEach(repository -> processRepository(internalAccessModel,
                                                                                             storage.getStorageId(),
                                                                                             repository)));
@@ -162,7 +160,6 @@ public class UsersConfig
         String key = "/storages/" + storageId + "/" + repository.getRepositoryId();
 
         repository.getPrivileges()
-                  .getPrivileges()
                   .forEach(privilege -> defaultPrivileges.add(privilege.getName().toUpperCase()));
 
         internalAccessModel.getRepositoryPrivileges().put(key, defaultPrivileges);
